+++ src/builtin/mod.rs
@@ -1,7 +1,9 @@
 mod idle;
+mod restart_wrapper;
 #[cfg(feature = "signal")]
 mod signal_receiver;
 
 pub use idle::IdleProcess;
+pub use restart_wrapper::{RestartBackoff, RestartWrapper};
 #[cfg(feature = "signal")]
 pub use signal_receiver::SignalReceiver;